+++ packages/rax-plugin-multi-pages/package.json
@@ -20,8 +20,8 @@
     "chalk": "^2.4.2",
     "console-clear": "^1.1.1",
     "fs-extra": "^8.1.0",
+    "handlebars": "^4.5.1",
     "loader-utils": "^1.1.0",
-    "qrcode-terminal": "^0.12.0",
     "rax-compile-config": "^0.2.0"
   }
 }
