+++ packages/rax-plugin-multi-pages/src/index.js
@@ -1,15 +1,13 @@
 const chalk = require('chalk');
-const path = require('path');
 const consoleClear = require('console-clear');
 const { handleWebpackErr } = require('rax-compile-config');
-const qrcode = require('qrcode-terminal');
 
 const getEntries = require('./getEntries');
 const setEntry = require('./setEntry');
+const setDevServer = require('./setDevServer');
 
 module.exports = ({ context, chainWebpack, onHook }) => {
-  const { rootDir, command, userConfig } = context;
-  const { outputDir } = userConfig;
+  const { command } = context;
   const entries = getEntries(context);
 
   let targets = [];
@@ -17,7 +15,16 @@ module.exports = ({ context, chainWebpack, onHook }) => {
   chainWebpack((config) => {
     targets = context.__configArr.map(v => v.name);
 
-    if (~targets.indexOf('web')) {
+    if (command === 'dev' && process.env.RAX_SSR !== 'true') {
+      setDevServer({
+        config,
+        context,
+        targets,
+        entries,
+      });
+    }
+
+    if (targets.includes('web')) {
       const webConfig = config.getConfig('web');
 
       setEntry(webConfig, context, entries, 'web');
@@ -28,32 +35,9 @@ module.exports = ({ context, chainWebpack, onHook }) => {
           isMultiple: true,
         }];
       });
-
-      if (command === 'dev' && process.env.RAX_SSR !== 'true') {
-        webConfig.devServer.set('before', (app, devServer) => {
-          const compiler = devServer.compiler.compilers[0];
-          const memFs = compiler.outputFileSystem;
-
-          entries.forEach(({ entryName }) => {
-            app.get(`/pages/${entryName}`, function(req, res) {
-              const htmlPath = path.resolve(rootDir, outputDir, `web/${entryName}.html`);
-
-              if (memFs.existsSync(htmlPath)) {
-                const outPut = memFs.readFileSync(htmlPath).toString();
-                res.send(outPut);
-              } else {
-                compiler.hooks.afterCompile.tap('sendHtml', () => {
-                  const outPut = memFs.readFileSync(htmlPath).toString();
-                  res.send(outPut);
-                });
-              }
-            });
-          });
-        });
-      }
     }
 
-    if (~targets.indexOf('weex')) {
+    if (targets.includes('weex')) {
       setEntry(config.getConfig('weex'), context, entries, 'weex');
     }
   });
@@ -65,33 +49,11 @@ module.exports = ({ context, chainWebpack, onHook }) => {
       return;
     }
 
-    console.log(chalk.green('Rax build finished:'));
-    console.log();
-
     console.log(chalk.green('Rax development server has been started:'));
     console.log();
 
-    if (~targets.indexOf('web')) {
-      console.log(chalk.green('[Web] Development server at:'));
-      entries.forEach(({ entryName }) => {
-        console.log('   ', chalk.underline.white(`${url}/pages/${entryName}`));
-      });
-      console.log();
-    }
-
-    if (~targets.indexOf('weex')) {
-      console.log(chalk.green('[Weex] Development server at:'));
-      entries.forEach(({ entryName }) => {
-        const weexUrl = `${url}/weex/${entryName}.js?wh_weex=true`;
-        console.log('   ', chalk.underline.white(weexUrl));
-        if (process.env.QRCODE === 'true') {
-          console.log();
-          qrcode.generate(weexUrl, {small: true});
-        }
-      });
-      console.log();
-
-      console.log(chalk.yellow('Set environment `QRCODE=true` to see weex qrcode.'));
-    }
+    console.log(chalk.green('Multi pages development server at:'));
+    console.log('   ', chalk.underline.white(url));
+    console.log();
   });
 };
