+++ packages/rax-plugin-multi-pages/src/template/main.hbs
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <title>Rax mutiple pages app</title>
+  <style>
+    .btn {
+      display: inline-block;
+      padding: 1px 7px;
+      border: 1px solid #ddd;
+      cursor: pointer;
+      text-decoration: none;
+      color: #000;
+      margin-top: 8px;
+      margin-right: 8px;
+    }
+    .btn:hover {
+      color: #f29718;
+    }
+  </style>
+</head>
+
+<body>
+  <h1>Rax mutiple pages app</h1>
+  {{#if entries.length}}
+  <h3>Page list</h3>
+  <ul>
+    {{#each entries}}
+    <li>
+      <div>/pages/{{entryName}}</div>
+      {{#if @root.hasWeb}}
+        <a class="btn" href="/pages/{{entryName}}" target="_blank">Web Preview</a>
+      {{/if}}
+      {{#if @root.hasWeex}}
+        <a class="btn" href="/weex/{{entryName}}.js?wh_weex=true" target="_blank">Weex Preview</a>
+      {{/if}}
+    </li>
+    {{/each}}
+  </ul>
+  {{/if}}
+</body>
+</html>
